refactor diffamp into its own circuit
